+++ static/javascript/content/android.js
@@ -8,6 +8,12 @@
 define(['Backbone'], function (Backbone) {
 	return Backbone.Collection.extend({
 		content: [
+			{
+				title: 'Android资源',
+				date: '2016-07-28',
+				url: '../static/notes/android/android-resource.html',
+				des: 'Android资源'
+			},
 		]
 	});
 });