+++ webpackVersion/scripts/content/android.js
@@ -1,5 +1,10 @@
 module.exports = {
 	title: 'android',
 	contents: [
+		{
+			title: 'Android资源',
+			date: '2016-07-28',
+			url: 'static/notes/android/android-resource.html'
+		}
 	]
 };